Crate binaural_beat_generator_cli

This program allows for the user to be able to listen to binaural beat tones on their machine.

Binaural beats is the idea that your brainwaves could be changed in a positive way by listening to slightly different tones in each ear.

Study suggests that binurial beats can help in stress relief, pain relief, anxiety relief as well as help in increased focus and so on.

See the following for more info. What Are Binaural Beats?

Modules

modules 🔒

A module that contains references related to all custom modules used.

Functions

main 🔒

This is the entry point to the program.

print_program_info 🔒

A helper function that just prints out the program name and author.

run_binaural_beat 🔒

A helper funciton that sets off the running of the binaural beat tones. It also spawns a new thread in order to watch for early completion.