binaural_beat_generator_cli::modules

Module preset

A module that contains code that allows for presets so that all settings can be easily used and passed around.

Structs

BinauralPresetGroup

This structure groups the basic values needed to run the binaural beat program.

Enums

Preset

The preset enum allows the user to be able to select a preset to use on the command line.

Functions

preset_list

This function returns all of the presets used in a vector.